qa/cephfs: upgrade caps_helper.CapTester.run_mon_cap_tests()
Design of run_mon_cap_tests() is now aligned with rest of CapTester --
it's not meant to inherited by test class anymore and it is to be run by
using CapTester instance. This commit also changes working of this
method, instead of obtaining FS from FS objects, it instead obtains
those names from MON cap.

Signed-off-by: Rishabh Dave <ridave@redhat.com>